@@ -57,11 +57,11 @@ add_action('admin_footer', function () {
             if ( event === 'form_saved' || event === 'form_save_email_requested' ) {
                 mergeTags['other'].tags.push( {
                     tag:  '{save_link}',
-                    label: <?php echo json_encode( esc_html__( 'Save & Continue Link', 'gravityforms' ) ); ?>
+                    label: <?php echo json_encode(esc_html__('Save & Continue Link', 'gravityforms')); ?>
                 } );
                 mergeTags['other'].tags.push( {
                     tag:   '{save_token}',
-                    label: <?php echo json_encode( esc_html__( 'Save & Continue Token', 'gravityforms' ) ); ?>
+                    label: <?php echo json_encode(esc_html__('Save & Continue Token', 'gravityforms')); ?>
                 } );
             }
             return mergeTags;
