@@ -11,53 +11,53 @@ use function is_string;
 
 abstract class Field implements Htmlable
 {
-  use Augment;
+    use Augment;
 
-  protected mixed $value;
-  protected mixed $raw;
-  protected string $additional_classes = "";
+    protected mixed $value;
+    protected mixed $raw;
+    protected string $additional_classes = '';
 
-  public function __construct(mixed $value)
-  {
-    $this->value = $value;
-    $this->raw = $value;
-  }
+    public function __construct(mixed $value)
+    {
+        $this->value = $value;
+        $this->raw = $value;
+    }
 
-  public static function getBlueprint(): Collection
-  {
-    return collect(static::blueprint());
-  }
+    public static function getBlueprint(): Collection
+    {
+        return collect(static::blueprint());
+    }
 
-  abstract protected static function blueprint(): array;
+    abstract protected static function blueprint(): array;
 
-  public function raw(): mixed
-  {
-    return $this->raw;
-  }
+    public function raw(): mixed
+    {
+        return $this->raw;
+    }
 
-  public function __toString(): string
-  {
-    $value = $this->value();
+    public function __toString(): string
+    {
+        $value = $this->value();
 
-    return match (true) {
-      is_string($value) => $value,
+        return match (true) {
+            is_string($value) => $value,
       default => "<!-- {$this->type()} could not output value as a string -->"
-    };
-  }
+        };
+    }
 
-  public function toHtml(): string
-  {
-    return $this->__toString();
-  }
+    public function toHtml(): string
+    {
+        return $this->__toString();
+    }
 
-  public function withClass(string $class): self
-  {
-    $this->additional_classes = $class;
-    return $this;
-  }
+    public function withClass(string $class): self
+    {
+        $this->additional_classes = $class;
+        return $this;
+    }
 
-  public function value(): mixed
-  {
-    return $this->augmented()->value;
-  }
+    public function value(): mixed
+    {
+        return $this->augmented()->value;
+    }
 }
