@@ -9,24 +9,24 @@ use function wp_get_attachment_image;
 
 class MediaField extends Field
 {
-  protected static function blueprint(): array
-  {
-    return [
+    protected static function blueprint(): array
+    {
+        return [
       'config' => [],
     ];
-  }
+    }
 
-  public function __toString(): string
-  {
-    return collect($this->value())
+    public function __toString(): string
+    {
+        return collect($this->value())
       ->map(fn ($item) => wp_get_attachment_image_url($item['id'], 'full'))
       ->first();
-  }
+    }
 
-  public function toHtml(): string
-  {
-    return collect($this->value())
+    public function toHtml(): string
+    {
+        return collect($this->value())
       ->map(fn ($item) => wp_get_attachment_image($item['id'], 'full', false, ['class' => $this->additional_classes]))
       ->implode('');
-  }
+    }
 }
