@@ -5,14 +5,12 @@ namespace Buildystrap\Builder\Fields;
 use Buildystrap\Builder\Extends\Field;
 
 use function collect;
-use function sprintf;
 
 class TitleField extends Field
 {
-
-  protected static function blueprint(): array
-  {
-    return [
+    protected static function blueprint(): array
+    {
+        return [
       'config' => [
         'label' => 'Heading',
         'tinymce' => [
@@ -27,35 +25,35 @@ class TitleField extends Field
         ],
       ],
     ];
-  }
+    }
 
-  public function augment(): void
-  {
-    $this->value = collect((array) $this->raw);
-  }
+    public function augment(): void
+    {
+        $this->value = collect((array) $this->raw);
+    }
 
-  public function titleClass(string $class): static
-  {
-    $this->additional_classes = $class;
-    return $this;
-  }
+    public function titleClass(string $class): static
+    {
+        $this->additional_classes = $class;
+        return $this;
+    }
 
-  public function __toString()
-  {
-    return $this->value()->get('text', '');
-  }
+    public function __toString()
+    {
+        return $this->value()->get('text', '');
+    }
 
-  public function toHtml(): string
-  {
-    return view('builder.components.title')->with([
-      "title" => [
-        "level" => $this->value()->get('level'),
-        "text" => $this->value()->get('text', ''),
-        "size" => $this->value()->get('size', []),
-        "color" => $this->value()->get('color', []),
-        "weight" => $this->value()->get('weight', []),
-        "class" => $this->value()->get('class', '') . ' ' . ($this->additional_classes ?? ''),
+    public function toHtml(): string
+    {
+        return view('builder.components.title')->with([
+      'title' => [
+        'level' => $this->value()->get('level'),
+        'text' => $this->value()->get('text', ''),
+        'size' => $this->value()->get('size', []),
+        'color' => $this->value()->get('color', []),
+        'weight' => $this->value()->get('weight', []),
+        'class' => $this->value()->get('class', '') . ' ' . ($this->additional_classes ?? ''),
       ]
     ])->render();
-  }
+    }
 }
