@@ -6,9 +6,9 @@ use Buildystrap\Builder\Extends\Module;
 
 class ButtonModule extends Module
 {
-  protected static function blueprint(): array
-  {
-    return [
+    protected static function blueprint(): array
+    {
+        return [
       'icon' => 'fa-solid fa-stop',
       'fields' => [
         'button' => [
@@ -22,5 +22,5 @@ class ButtonModule extends Module
         'selectorTab' => 'regular'
       ]
     ];
-  }
+    }
 }
