@@ -11,18 +11,18 @@ use Buildystrap\ThemeOptions;
 :root {
 <?php
 
-if (!function_exists('generateStructureVars')) {
-  add_action('generate-structure-vars', [ThemeOptions::class, 'generateStructureVars']);
+if ( ! function_exists('generateStructureVars')) {
+    add_action('generate-structure-vars', [ThemeOptions::class, 'generateStructureVars']);
 }
 do_action('generate-structure-vars');
 
-if (!function_exists('generateColorVars')) {
-  add_action('generate-color-vars', [ThemeOptions::class, 'generateColorVars']);
+if ( ! function_exists('generateColorVars')) {
+    add_action('generate-color-vars', [ThemeOptions::class, 'generateColorVars']);
 }
 do_action('generate-color-vars');
 
-if (!function_exists('generateTypographyVars')) {
-  add_action('generate-typography-vars', [ThemeOptions::class, 'generateTypographyVars']);
+if ( ! function_exists('generateTypographyVars')) {
+    add_action('generate-typography-vars', [ThemeOptions::class, 'generateTypographyVars']);
 }
 do_action('generate-typography-vars');
 ?>
@@ -30,7 +30,7 @@ do_action('generate-typography-vars');
 
 /******* Utils */
 <?php
-if (!function_exists('generateColorUtils')) {
-  add_action('generate-color-utils', [ThemeOptions::class, 'generateColorUtils']);
+if ( ! function_exists('generateColorUtils')) {
+    add_action('generate-color-utils', [ThemeOptions::class, 'generateColorUtils']);
 }
 do_action('generate-color-utils');
