@@ -15,28 +15,27 @@
  * @version 7.0.1
  */
 
-defined( 'ABSPATH' ) || exit;
+defined('ABSPATH') || exit;
 
 /*
  * @hooked wc_empty_cart_message - 10
  */
-do_action( 'woocommerce_cart_is_empty' );
+do_action('woocommerce_cart_is_empty');
 
-if ( wc_get_page_id( 'shop' ) > 0 ) :
-	?>
+if (wc_get_page_id('shop') > 0) :
+    ?>
 	<p class="return-to-shop">
-		<a class="btn btn-outline-primary" href="<?php echo esc_url( apply_filters( 'woocommerce_return_to_shop_redirect', wc_get_page_permalink( 'shop' ) ) ); ?>">
+		<a class="btn btn-outline-primary" href="<?php echo esc_url(apply_filters('woocommerce_return_to_shop_redirect', wc_get_page_permalink('shop'))); ?>">
 			<?php
-				/**
-				 * Filter "Return To Shop" text.
-				 *
-				 * @since 4.6.0
-				 * @param string $default_text Default text.
-				 */
-				echo esc_html( apply_filters( 'woocommerce_return_to_shop_text', __( 'Return to shop', 'woocommerce' ) ) );
-			?>
+                /**
+                 * Filter "Return To Shop" text.
+                 *
+                 * @since 4.6.0
+                 * @param string $default_text Default text.
+                 */
+                echo esc_html(apply_filters('woocommerce_return_to_shop_text', __('Return to shop', 'woocommerce')));
+            ?>
 		</a>
 	</p>
 	<?php
 endif;
-
