@@ -15,7 +15,7 @@
  * @version 7.0.1
  */
 
-defined( 'ABSPATH' ) || exit;
+defined('ABSPATH') || exit;
 
 $totals = $order->get_order_item_totals(); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
 ?>
@@ -24,43 +24,43 @@ $totals = $order->get_order_item_totals(); // phpcs:ignore WordPress.WP.GlobalVa
 	<table class="shop_table">
 		<thead>
 			<tr>
-				<th class="product-name"><?php esc_html_e( 'Product', 'woocommerce' ); ?></th>
-				<th class="product-quantity"><?php esc_html_e( 'Qty', 'woocommerce' ); ?></th>
-				<th class="product-total"><?php esc_html_e( 'Totals', 'woocommerce' ); ?></th>
+				<th class="product-name"><?php esc_html_e('Product', 'woocommerce'); ?></th>
+				<th class="product-quantity"><?php esc_html_e('Qty', 'woocommerce'); ?></th>
+				<th class="product-total"><?php esc_html_e('Totals', 'woocommerce'); ?></th>
 			</tr>
 		</thead>
 		<tbody>
-			<?php if ( count( $order->get_items() ) > 0 ) : ?>
-				<?php foreach ( $order->get_items() as $item_id => $item ) : ?>
+			<?php if (count($order->get_items()) > 0) : ?>
+				<?php foreach ($order->get_items() as $item_id => $item) : ?>
 					<?php
-					if ( ! apply_filters( 'woocommerce_order_item_visible', true, $item ) ) {
-						continue;
-					}
-					?>
-					<tr class="<?php echo esc_attr( apply_filters( 'woocommerce_order_item_class', 'order_item', $item, $order ) ); ?>">
+                    if ( ! apply_filters('woocommerce_order_item_visible', true, $item)) {
+                        continue;
+                    }
+                    ?>
+					<tr class="<?php echo esc_attr(apply_filters('woocommerce_order_item_class', 'order_item', $item, $order)); ?>">
 						<td class="product-name">
 							<?php
-								echo wp_kses_post( apply_filters( 'woocommerce_order_item_name', $item->get_name(), $item, false ) );
+                                echo wp_kses_post(apply_filters('woocommerce_order_item_name', $item->get_name(), $item, false));
 
-								do_action( 'woocommerce_order_item_meta_start', $item_id, $item, $order, false );
+                                do_action('woocommerce_order_item_meta_start', $item_id, $item, $order, false);
 
-								wc_display_item_meta( $item );
+                                wc_display_item_meta($item);
 
-								do_action( 'woocommerce_order_item_meta_end', $item_id, $item, $order, false );
-							?>
+                                do_action('woocommerce_order_item_meta_end', $item_id, $item, $order, false);
+                            ?>
 						</td>
-						<td class="product-quantity"><?php echo apply_filters( 'woocommerce_order_item_quantity_html', ' <strong class="product-quantity">' . sprintf( '&times;&nbsp;%s', esc_html( $item->get_quantity() ) ) . '</strong>', $item ); ?></td><?php // @codingStandardsIgnoreLine ?>
-						<td class="product-subtotal"><?php echo $order->get_formatted_line_subtotal( $item ); ?></td><?php // @codingStandardsIgnoreLine ?>
+						<td class="product-quantity"><?php echo apply_filters('woocommerce_order_item_quantity_html', ' <strong class="product-quantity">' . sprintf('&times;&nbsp;%s', esc_html($item->get_quantity())) . '</strong>', $item); ?></td><?php // @codingStandardsIgnoreLine?>
+						<td class="product-subtotal"><?php echo $order->get_formatted_line_subtotal($item); ?></td><?php // @codingStandardsIgnoreLine?>
 					</tr>
 				<?php endforeach; ?>
 			<?php endif; ?>
 		</tbody>
 		<tfoot>
-			<?php if ( $totals ) : ?>
-				<?php foreach ( $totals as $total ) : ?>
+			<?php if ($totals) : ?>
+				<?php foreach ($totals as $total) : ?>
 					<tr>
-						<th scope="row" colspan="2"><?php echo $total['label']; ?></th><?php // @codingStandardsIgnoreLine ?>
-						<td class="product-total"><?php echo $total['value']; ?></td><?php // @codingStandardsIgnoreLine ?>
+						<th scope="row" colspan="2"><?php echo $total['label']; ?></th><?php // @codingStandardsIgnoreLine?>
+						<td class="product-total"><?php echo $total['value']; ?></td><?php // @codingStandardsIgnoreLine?>
 					</tr>
 				<?php endforeach; ?>
 			<?php endif; ?>
@@ -68,31 +68,31 @@ $totals = $order->get_order_item_totals(); // phpcs:ignore WordPress.WP.GlobalVa
 	</table>
 
 	<div id="payment">
-		<?php if ( $order->needs_payment() ) : ?>
+		<?php if ($order->needs_payment()) : ?>
 			<ul class="wc_payment_methods payment_methods methods">
 				<?php
-				if ( ! empty( $available_gateways ) ) {
-					foreach ( $available_gateways as $gateway ) {
-						wc_get_template( 'checkout/payment-method.php', array( 'gateway' => $gateway ) );
-					}
-				} else {
-					echo '<li class="woocommerce-notice woocommerce-notice--info woocommerce-info">' . apply_filters( 'woocommerce_no_available_payment_methods_message', esc_html__( 'Sorry, it seems that there are no available payment methods for your location. Please contact us if you require assistance or wish to make alternate arrangements.', 'woocommerce' ) ) . '</li>'; // @codingStandardsIgnoreLine
-				}
-				?>
+                if ( ! empty($available_gateways)) {
+                    foreach ($available_gateways as $gateway) {
+                        wc_get_template('checkout/payment-method.php', [ 'gateway' => $gateway ]);
+                    }
+                } else {
+                    echo '<li class="woocommerce-notice woocommerce-notice--info woocommerce-info">' . apply_filters('woocommerce_no_available_payment_methods_message', esc_html__('Sorry, it seems that there are no available payment methods for your location. Please contact us if you require assistance or wish to make alternate arrangements.', 'woocommerce')) . '</li>'; // @codingStandardsIgnoreLine
+                }
+                ?>
 			</ul>
 		<?php endif; ?>
 		<div class="form-row">
 			<input type="hidden" name="woocommerce_pay" value="1" />
 
-			<?php wc_get_template( 'checkout/terms.php' ); ?>
+			<?php wc_get_template('checkout/terms.php'); ?>
 
-			<?php do_action( 'woocommerce_pay_order_before_submit' ); ?>
+			<?php do_action('woocommerce_pay_order_before_submit'); ?>
 
-			<?php echo apply_filters( 'woocommerce_pay_order_button_html', '<button type="submit" class="btn btn-primary" id="place_order" value="' . esc_attr( $order_button_text ) . '" data-value="' . esc_attr( $order_button_text ) . '">' . esc_html( $order_button_text ) . '</button>' ); // @codingStandardsIgnoreLine ?>
+			<?php echo apply_filters('woocommerce_pay_order_button_html', '<button type="submit" class="btn btn-primary" id="place_order" value="' . esc_attr($order_button_text) . '" data-value="' . esc_attr($order_button_text) . '">' . esc_html($order_button_text) . '</button>'); // @codingStandardsIgnoreLine?>
 
-			<?php do_action( 'woocommerce_pay_order_after_submit' ); ?>
+			<?php do_action('woocommerce_pay_order_after_submit'); ?>
 
-			<?php wp_nonce_field( 'woocommerce-pay', 'woocommerce-pay-nonce' ); ?>
+			<?php wp_nonce_field('woocommerce-pay', 'woocommerce-pay-nonce'); ?>
 		</div>
 	</div>
 </form>
