@@ -17,31 +17,31 @@
  * @version 3.2.0
  */
 
-if ( ! defined( 'ABSPATH' ) ) {
-	exit;
+if ( ! defined('ABSPATH')) {
+    exit;
 }
 
 $downloads     = WC()->customer->get_downloadable_products();
 $has_downloads = (bool) $downloads;
 
-do_action( 'woocommerce_before_account_downloads', $has_downloads ); ?>
+do_action('woocommerce_before_account_downloads', $has_downloads); ?>
 
-<?php if ( $has_downloads ) : ?>
+<?php if ($has_downloads) : ?>
 
-	<?php do_action( 'woocommerce_before_available_downloads' ); ?>
+	<?php do_action('woocommerce_before_available_downloads'); ?>
 
-	<?php do_action( 'woocommerce_available_downloads', $downloads ); ?>
+	<?php do_action('woocommerce_available_downloads', $downloads); ?>
 
-	<?php do_action( 'woocommerce_after_available_downloads' ); ?>
+	<?php do_action('woocommerce_after_available_downloads'); ?>
 
 <?php else : ?>
 	<div class="woocommerce-Message woocommerce-Message--info woocommerce-info">
-		<a class="btn btn-outline-primary" href="<?php echo esc_url( apply_filters( 'woocommerce_return_to_shop_redirect', wc_get_page_permalink( 'shop' ) ) ); ?>">
-			<?php esc_html_e( 'Browse products', 'woocommerce' ); ?>
+		<a class="btn btn-outline-primary" href="<?php echo esc_url(apply_filters('woocommerce_return_to_shop_redirect', wc_get_page_permalink('shop'))); ?>">
+			<?php esc_html_e('Browse products', 'woocommerce'); ?>
 		</a>
-		<?php esc_html_e( 'No downloads available yet.', 'woocommerce' ); ?>
+		<?php esc_html_e('No downloads available yet.', 'woocommerce'); ?>
 	</div>
 <?php endif; ?>
 
 <?php
-do_action( 'woocommerce_after_account_downloads', $has_downloads );
+do_action('woocommerce_after_account_downloads', $has_downloads);
